rename src/ir to src/bc
Calling this directory "ir" is confusing, since our IR is PIR, and RIR
is our bytecode. So I proposed to rename this to "bc".
