+++ rir/src/bc/Compiler.cpp
