+++ rir/src/compiler/backend.cpp
@@ -1,6 +1,8 @@
 #include "backend.h"
 #include "R/BuiltinIds.h"
 #include "analysis/dead.h"
+#include "bc/CodeStream.h"
+#include "bc/CodeVerifier.h"
 #include "compiler/analysis/cfg.h"
 #include "compiler/analysis/last_env.h"
 #include "compiler/analysis/reference_count.h"
@@ -13,8 +15,6 @@
 #include "compiler/util/lowering/allocators.h"
 #include "compiler/util/visitor.h"
 #include "interpreter/instance.h"
-#include "ir/CodeStream.h"
-#include "ir/CodeVerifier.h"
 #include "runtime/DispatchTable.h"
 #include "simple_instruction_list.h"
 #include "utils/FunctionWriter.h"
