+++ rir/src/compiler/pir/deopt_context.h
@@ -1,8 +1,8 @@
 #pragma once
 
+#include "bc/BC_inc.h"
 #include "common.h"
 #include "continuation_context.h"
-#include "ir/BC_inc.h"
 #include "runtime/LazyEnvironment.h"
 
 #include <vector>
