+++ rir/src/compiler/rir2pir/rir2pir.cpp
@@ -3,6 +3,8 @@
 #include "R/Funtab.h"
 #include "R/RList.h"
 #include "R/Symbols.h"
+#include "bc/BC.h"
+#include "bc/Compiler.h"
 #include "compiler/analysis/cfg.h"
 #include "compiler/analysis/query.h"
 #include "compiler/analysis/verifier.h"
@@ -14,8 +16,6 @@
 #include "compiler/util/bb_transform.h"
 #include "compiler/util/visitor.h"
 #include "insert_cast.h"
-#include "ir/BC.h"
-#include "ir/Compiler.h"
 #include "runtime/ArglistOrder.h"
 #include "simple_instruction_list.h"
 #include "utils/FormalArgs.h"
