+++ rir/src/runtime/Deoptimization.cpp
